fix(macos): make the System Extension actually build (M2) — cycle, FFI types, entry point
Building the M2-M6 scaffolding on a real Mac (Xcode 26.6, Apple Silicon) surfaced three bugs it
shipped with; all fixed, and the SE now compiles + links the Rust engine into an arm64 Mach-O
executable (nil_start/nil_stop/nil_ingest_packets/nil_negotiated_mtu defined in the binary):

- Build-graph cycle: the SE target both LINKED NilApple.xcframework and GENERATED it via an
  in-target script phase ("Cycle inside PacketTunnel"). Move the engine build out-of-band into
  client/apple/build-engine.sh (cargo -> universal static lib -> xcframework); the target now only
  links the pre-built artifact. project.yml drops the script phase.
- FFI type mismatches in the shared PacketTunnelProvider.swift (would have broken iOS too):
  Data(bytes:count:) needs Int but the FFI len is usize->UInt; nil_ingest_packets' lens is
  *const usize->UInt while d.count is Int. Fixed both casts.
- Missing macOS entry point: a NetworkExtension *system extension* is a standalone executable and
  needs _main (the iOS appex uses NSExtensionPrincipalClass instead). Add PacketTunnelMain.swift
  calling NEProvider.startSystemExtensionMode(); wired into the SE target sources.

Verified locally (arm64, unsigned). gitignore the xcodebuild build/ dir.

M3 (activate + connect) remains BLOCKED on a paid Apple Developer Program: a free/personal team
cannot provision the Network Extension capability (signed build fails "No profiles for
'com.nilvpn.client.PacketTunnel' were found"). MACOS_DEVICE_VERIFY.md updated with the verified
build status and the account gate.

+++ crates/nil-apple/apple/PacketTunnelProvider.swift
@@ -32,7 +32,7 @@ final class PacketTunnelProvider: NEPacketTunnelProvider {
         let writeCb: NilWriteCb = { ctx, pkt, len, af in
             let me = Unmanaged<PacketTunnelProvider>.fromOpaque(ctx!).takeUnretainedValue()
             let proto = NSNumber(value: af == 30 ? AF_INET6 : AF_INET)
-            me.packetFlow.writePackets([Data(bytes: pkt!, count: len)], withProtocols: [proto])
+            me.packetFlow.writePackets([Data(bytes: pkt!, count: Int(len))], withProtocols: [proto])
         }
         let statusCb: NilStatusCb = { ctx, state, _ in
             let me = Unmanaged<PacketTunnelProvider>.fromOpaque(ctx!).takeUnretainedValue()
@@ -72,7 +72,7 @@ final class PacketTunnelProvider: NEPacketTunnelProvider {
             for d in datas {
                 d.withUnsafeBytes { raw in
                     var p = raw.bindMemory(to: UInt8.self).baseAddress
-                    var len = d.count
+                    var len = UInt(d.count)   // FFI lens is *const usize → UInt
                     var af: Int32 = 0
                     nil_ingest_packets(t, &p, &len, &af, 1)   // engine copies synchronously
                 }
